+++ CHANGES.md
@@ -39,6 +39,7 @@
 - Rename setting `magik.typing.showAtomInlayHints` to `magik.typing.showTypingInlayHints`.
 - Replace `magik.libsDirs` with `magik.productDirs`, register products instead of libs directories.
 - Dump types database (per workspace folder) on shutdown, read it on start up and re-index any changed files, for quicker initialization.
+- Transmit a product.def/module.def/load_list.txt with F4-b from VSCode.
 - Several fixes.
 
 0.9.1 (2024-03-13)